refactor: 공용 컴포넌트 shared 폴더 파일정리
